Refactor out the binary blob delivery and move to deploy from source
This currently is not configurable - and statically deploys v0.4.0
release of flannel. I would ideally like to make this a bit more
configurable in terms of version, and location to isntall from + provide
an option to fetch / share the compiled binary.

first iteration landing of the code though, tally ho