Updated the etcd_rest.py script with a try/catch to prevent failure and locking deployments in CI
Amulet has a 90% accuracy rating when determining if the environment has
fully settled. There are cases where the test starts before the env has
fully settled and the ETCD unit is torn down before flannel has a chance
ot relate to the etcd node. The test isn't checking against
flannel-docker/1 which is the node in question thats failing. This
try/except block allows the unit to enter a blocked state and not fail
the deployment test.

This will undoutebly surface other cases where failure lands the flannel
unit in an inconsistent state and will need to be coupled with an action
to recover the service if ETCD registration fails.